Index(Home)
Se elimina el array, se agrega en enlace para que se conecte al archivo data.js y se crea el carousel de tarjetas para que quede igual que en el archivo past.
