+++ index.html
@@ -5,17 +5,17 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="shortcut icon" href="./images/Favicon.png" type="image/x-icon">
+    <link rel="shortcut icon" href="../images/Favicon.png" type="image/x-icon">
     <title>Home</title>
-    <!--style-->
-    <link rel="stylesheet" href="./style/style.css">
+
     <!--bootstrap -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css">
     <!--bootstrap icons-->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.3/font/bootstrap-icons.css">
-
-      <!-- Link Swiper's CSS -->
-      <link rel="stylesheet" href="https://unpkg.com/swiper/swiper-bundle.min.css" />
+    <!-- swiper -->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@9/swiper-bundle.min.css">
+    <!--style - css-->
+    <link rel="stylesheet" href="../style/style.css">
 </head>
 
 <body>
@@ -97,16 +97,18 @@
         </form>
     </div>
 
-    <!--cards  -->
-    <div class="container">
+    <!--cards-->
+    <div class="container-carousel">
         <div class="swiper-container mySwiper">
-            <div class="swiper-wrapper">
-                <div   id="eventosContainer" class=" row d-flex justify-content-around " >
-                    
-                </div>
+            <div id="section" class="swiper-wrapper">
+                
             </div>
+            <div class=" btn-swiper swiper-button-next"></div>
+            <div class=" btn-swiper swiper-button-prev"></div>
         </div>
     </div>
+
+     
     <!--footer-->
     <footer class=" redes d-flex justify-content-between">
         <div class="  d-flex justify-content-start ">
@@ -119,13 +121,18 @@
                 colombia </h1>
         </div>
     </footer>
+    <!-- boostrap -->
     <script src="	https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
+    <!-- swiper -->
+    <script src="https://cdn.jsdelivr.net/npm/swiper@9/swiper-bundle.min.js"></script>
+    <script src="https://unpkg.com/swiper/swiper-bundle.min.js"></script>
+    <!-- Initialize Swiper -->
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js"
+        integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3"
+        crossorigin="anonymous"></script>
 
-     <!-- Swiper JS -->
-
-     <script src="https://unpkg.com/swiper/swiper-bundle.min.js"></script>
-     <script src="./index.js" ></script>
-    
+    <!--javaScript-->
+    <script src="./javaScript/index.js" type="module"></script>
 </body>
 
 </html>
