+++ style/style.css
@@ -259,32 +259,3 @@ margin-top: 22%;
 .btn-swiper {
     color: #000;
 }
-
-/* Carousel Home */
-
- .container{
-   width: 90%;
- 
-   padding: 10px 40px;
- }
- .container .swiper-container{
-   width: 100%;
-   height: 100%;
- }
- .container .swiper-slide {
-   position: relative; 
-   text-align: center;
-   font-size: 18px;
-   background: #fff;
-   border-radius: 15px;
-   text-align: start;
-   box-shadow: 0px 4px 4px rgba(196, 196, 196, 0.25);
-  
- }
-
- .swiper-button-next{
-     right: 26px !important;
- }
- .swiper-button-prev{
-     left: 26px !important;
- }
